+++ changelog.md
@@ -1,6 +1,8 @@
 ## Current
 - Updated to support new args approach for executors.
-
+- Fixed Android execution for Kodi Nexus.
+- Android storage issue fixes
+  
 ## Previous releases
 - Updated documentation.
 - Optimized addon settings.